Add Privacy Policy and Terms & Conditions pages
- /privacy-policy and /terms-and-conditions, content in src/data/legal.ts.
  The privacy policy describes what this site actually does: form data, the
  services that process it (Vercel, Neon, Resend, Cloudflare Turnstile,
  Upstash, Google Maps embeds), no tracking cookies, retention, overseas
  storage and how to request access/correction/deletion. The terms cover
  services, no guaranteed admission/visa/score outcomes, applicant duties,
  fees/refunds, acceptable use, IP and Bangladeshi governing law.
  DRAFT: marked in the file for GlobalEd management / legal review.
- Footer links now point to them ('Terms of Use' renamed 'Terms &
  Conditions'); both added to the sitemap at low priority.

+++ src/lib/validation/dates.ts
@@ -39,11 +39,11 @@ export function isRealDate(value: string): boolean {
   return !Number.isNaN(d.getTime()) && d.toISOString().slice(0, 10) === value;
 }
 
-/** "2026-09-23" -> "23 Sep 2026" for messages. */
+/** "2026-09-23" -> "23 September 2026". */
 export function formatDate(isoDate: string): string {
   return new Date(`${isoDate}T00:00:00Z`).toLocaleDateString("en-GB", {
     day: "numeric",
-    month: "short",
+    month: "long",
     year: "numeric",
     timeZone: "UTC",
   });
